add test for updating asset instead of creating a new one
