+++ synq/video2.go
@@ -64,20 +64,36 @@ func (v VideoV2) GetAsset(assetId string) (Asset, error) {
 	return asset, err
 }
 
-func (v VideoV2) CreateOrUpdateAsset(asset *Asset) error {
-	// make sure the API is set
-	asset.Api = *v.Api
-	url := v.Api.getBaseUrl() + "/assets"
-	data, _ := json.Marshal(asset)
-	body := bytes.NewBuffer(data)
-	err := asset.handleAssetReq("POST", url, body)
-	if err == nil {
-		v.Assets = append(v.Assets, *asset)
+func (v *VideoV2) FindAsset(location string) (Asset, bool) {
+	for _, a := range v.Assets {
+		if a.Location == location && a.Id != "" {
+			return a, true
+		}
 	}
-	return err
+	return Asset{}, false
 }
 
-func (v VideoV2) CreateAsset(state, fileType, location string) (Asset, error) {
+func (v *VideoV2) CreateOrUpdateAsset(asset *Asset) error {
+	// make sure the API is set
+	asset.Api = *v.Api
+	// check if this asset exists, if it does, just update
+	a, found := v.FindAsset(asset.Location)
+	if found {
+		asset.Id = a.Id
+		return asset.Update()
+	} else {
+		url := v.Api.getBaseUrl() + "/assets"
+		data, _ := json.Marshal(asset)
+		body := bytes.NewBuffer(data)
+		err := asset.handleAssetReq("POST", url, body)
+		if err == nil {
+			v.Assets = append(v.Assets, *asset)
+		}
+		return err
+	}
+}
+
+func (v *VideoV2) CreateAsset(state, fileType, location string) (Asset, error) {
 	var asset Asset
 	asset.VideoId = v.Id
 	asset.State = state
